backend/requirements.txt
flask
flask-cors
twilio
werkzeug<3.0.0
flasgger
google-generativeai
python-dotenv
